Fix: QY of 2-hexanal
In SR mk03 in rxn_ket.f, an if statement has been added on l. 522, which
sets qy to 0, if yg1 is 0. Otherwise, qy will be NaN do to a devision by
yg1 in the calculation of eta.
Removal of idum. Wavelengths are read in directly as reals rather than
integer idum with conversion to reals.