Fix: 2-hexanal qy
In SR mk03 in rxn_ket.f, avoid division by 0 in the calculation of the
quantum yields for 2-hexanal. Deleting variable idum. Now, all quantum yields are read in as reals and not as integers with later conversion to 
reals.
Updating include file for the original mechanism coupling to DSMACC.
Removing unnecessary print commands in rdinp.f and TUV521.f.
Creating a gitignore for .o files.